backend/requirements.txt
Flask==3.1.0
python-dotenv~=1.0.1
pymongo
pandas
scipy
scikit-learn
mlxtend==0.23.1